pyproject.toml
[project]
name = "gnosari-engine"
version = "0.1.4"
description = "A powerful framework for orchestrating multi-agent teams using Large Language Models. Create intelligent AI agent swarms that collaborate through streaming delegation and dynamic tool discovery."
authors = [
    {name = "daviunx",email = "david.marsa@neomanex.com"}
]
readme = "README.md"
requires-python = ">=3.12,<=3.13.2"
dependencies = [
    "openai (>=1.100.1,<2.0.0)",
    "instructor (>=1.0.0,<2.0.0)",
    "python-dotenv (>=1.0.0,<2.0.0)",
    "rich (>=13.0.0,<14.0.0)",
    "PyYAML (>=6.0,<7.0)",
    "requests (>=2.31.0,<3.0.0)",
    "aiohttp (>=3.9.0,<4.0.0)",
    "tiktoken (>=0.5.0,<1.0.0)",
    "fastmcp (>=2.11.3,<3.0.0)",
    "pytest-asyncio (>=1.1.0,<2.0.0)",
    "lxml (>=6.0.1,<7.0.0)",
    "openai-agents (>=0.2.10,<0.3.0)",
    "sqlalchemy (>=2.0.0,<3.0.0)",
    "pymysql (>=1.1.0,<2.0.0)",
    "embedchain[youtube] (>=0.1.128,<0.2.0)",
    "youtube-transcript-api (>=1.2.2,<2.0.0)",
    "pytube (>=15.0.0,<16.0.0)",
    "aiosqlite (>=0.21.0,<0.22.0)",
    "greenlet (>=3.2.4,<4.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "celery[redis] (>=5.5.0,<6.0.0)",
    "redis (>=5.2.0,<6.0.0)",
    "flower (>=2.0.0,<3.0.0)",
    "psutil (>=6.1.0,<7.0.0)"
]

[project.scripts]
gnosari = "gnosari.cli:main"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
# Enable packaging and declare the package location using a src layout
package-mode = true
packages = [
    { include = "gnosari", from = "src" }
]

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"

